Added asint command, and modified asfloat mechanism.
The asint and asfloat commands will now behave similarly to the builtin
float and int commands, except that they do not raise ValueErrors for
invalid string input.  This means that you are guaranteed to get either
int or str types from asint, or float or int types from asfloat.

This is *slightly* slower than the original implementation, but it the
fast_* versions can be used if speed is needed.