Added optional string-centric arguments to functions.
For all conversion functions, a new "raise_on_invalid" optional argument
has been included that will cause a "ValueError" to be raised if the
input string cannot be converted to a number... this makes the functions
mimic the behavior of the built-in "float" or "int" functions instead of
simply returning the string as-is.

For the test functions, a "str_only" optional argument has been added
that will make the functions return False for any non-string input.

Docstrings and unit-tests have been updated to reflect these changes.