Reworked all fast_* to include overflow handling.
Each of the fast_* functions now do overflow detection, and fall
back on the Python-based string reading functions if it is determined
that there is overflow or precision is being lost.

This was done by modifying the fast_atoi and fast_atof functions to
do basic overflow checking, which includes checking if the value becomes
negative before the sign is applied, or if the exponent part is larger
than a known value that can lose precision. To improve precision and
speed, fast_atof now reads the numbers using integer arithmetic, then
performs a single FLOP to convert the number to floating point. Further,
the internal floating point operations are done using a long double,
which retains as much precision as possible before returning as double.

As part of the improvement for fast_atoi and fast_atof, support for
long literals (ints ending in 'l' or 'L') has been added for Python2.

The intent is that the safe_* functions will be obsoleted in favor
of the fast_* functions now that both are guaranteed to give full
precision yet fast_* will do it much more quickly for the most common
use cases (non-extreme numbers).

Timing tests indicate that for the extreme cases the fast_* functions
are still comparable for the safe_* functions.