Fixed segfault when the input is unicode.
When dealing with unicode input, the python object needs to be
converted to a bytes object before being converted to a character
array. Previously, fastnumbers was relying on the python object
remaining in memory when dealing with character arrays because a
strcpy was not performed.  Because extracting the character array from
unicode requires a temporary python object which is quickly
de-referenced, this is not a safe technique; the segfault is rare
because python garbage collects de-references objects only periodically,
so the character array typically remains in memory.

To solve this issue, all character arrays are now explicitly copied with
strcpy. This required modification of the conversion functions to
free the character array memory before returning from fastnumbers.

This resolves issue #2, and most likely resolves issue #1.