Fixed bug where strings with middle NUL chars were treated as valid.
Python strings can have NUL characters, so when converted to C strings
it can be a problem determining where the string should end. The
string to float/int functions were stopping halfway through strings
that began with numbers, had a NUL in the middle, then had other junk
at the end; as a result, they considered these invalid strings to
be valid.

Now, the expected length of the string according to Python is passed
to the conversion functions, and they make sure that only NUL whitespace
or NUL characters trail the actual numbers.