Fixed odd segfault in Python 3.5.
On Python 3.5 a segfault would occur when stripping whitespace from
unicode data.  This was done by calling the "strip" method of the string
object via 'PyObject_CallMethod(input, "strip", NULL)'.  On Python 3.4
and before this worked without issue, but on Python 3.5 it would return
NULL.

Investigation showed that it would output the following traceback:
"SystemError: ... returned a result with an error set". It seems that
there was a change to PyObject_Call in Python 3.5 and the change
was related to this type of error... please visit
https://bugs.python.org/issue23571.

To fix, a regular C loop was implemented that checks each unicode
character individually and ignores whitespace.