Fixed bug caused by uncaught errors.
Some python functions may return NULL, but the C code was not catching
this in all places. Further, if this happens but we want to ignore the
error, the error stack needs to be cleared (this became an exception
in Python 3.5).

The C code has become much more defensive and checks for errors in
more places. Additionally, the condition that exposes this issue has
been added to the unit tests.