Updated code to silence GCC compiler warnings.
Some macros returned a value but the caller was not using the value, so
explicit calls to (void) have been added.  Also, some functions that
are made up a switch statement with returns were complaining that
it was possible to get to the end without a return (even though clang
was able to identify that this was not possible).