Fixed annoying "hypothesis" errors.
The hypothesis module that is used for testing has a constantly changing
API that seems to introduce backwards-incompatible changes with every
update. The most recent one was a "FailedHealthCheck" error that
had to be eliminated by passing in random values through the function
arguments rather than calling directly within the function.