Fixed SystemError for dumb input.
Previously, if the user gave a class like the following:

class Test:
    def __float__(self):
        raise ValueError("something here might go wrong")

A SystemError would be raised on Python 3.5 because the exception
would be uncaught by fastnumbers.  This has been updated to do the right
thing.

Really, fastnumbers was never intended to handle this type of input,
and it's not clear why a class would be set up like this in the
first place. It doesn't lose too much performance for this small
increase in flexibility/safety, so maybe it's OK. I just hope this
doesn't begin a rabbit hole of scope creep where eventually I will be
converting lists and datetime objects to floats and ints...