Fixed unicode errors on windows with Python >= 3.3.
The Python API documentation implies that characters should be of Py_UNICODE type and
not Py_UCS4 for functions like Py_UNICODE_ISSPACE and Py_UNICODE_TONUMERIC.
Looking at the actual source code it turns out this is not correct, but rather Py_UCS4 should
be used. Previously, a cast to Py_UNICODE was being performed, but this cast has been
removed as it was causing incorrect results.