Fixed segfault in unicode handling on Python 3.6.
If a non-ASCII string was passed to fastnumbers that had leading
whitespace, fastnumbers would segfault saying it was trying to free
unallocated memory.

This has been solved by rewriting the function to use python
builtin objects. This is one of those rare cases when fixing a bug
resulted in cleaner looking code as well.