Parsing validity is now checked by string length.
The end pointer of the string is passed to the contains and parsing
functions, and when the operation is complete it is checked that
the string pointer is the same as the end pointer.

This is instead of the previous method of checking the trailing
characters after parsing are whitespace or NUL.

This method allows the removal of a strlen call that was used to check
that no NUL characters were contained in the character array. Now,
the end pointer is taken from Python's methods, so the end
position is known and cannot be mistaken by accidental NUL characters
in the string.