Improved string to int conversion speed.
The new method completely avoids overflow checking during the conversion
by checking the length of the string and skipping to the Python
implementation if it is over the largest length that can be converted
to a long without overflow.

Additionally, some heuristics have been added to quickly asses if
a string is likely an int, and quit early if not.