Improved string to float conversion speed.
Roughly the same methodology as was used for ints in the previous
commit has been used for floats here. Overflow checking is avoided
during parsing by using heuristics to skip to the Python parser
if it is detected that overflow is likely. Simple heuristics have been
added to quickly asses if a string is likely a float and quit early if
not.