Added an Options struct to pass options around.
This reduces the number of parameters each function has to accept,
and will make function modifications easier going forward because
only the Option struct will need to be changed.

Coupled with this is a modification of the error handling mechanism.
Previously, all errors were cleared when encountered, and then
the error was assigned only immediately before returning. Now,
errors are kept and to ensure that the real exceptions are raised,
and the hope is that any error created by fastnumbers itself will
match a built-in error.