Improve unicode translation performance.
If a unicode object is stored internally like ASCII, it can be
cast to a char* array directly instead of allocating new memory.
This is a significant time saver.

In cases where the internal unicode data is not already stored in
ASCII form, the speed of the for loop to perform the conversion has
been improved. This has been done by putting the simplest conditions
to test at the top of the if statement, and also by applying the
register modifier to the looping variables.