Fix out-of-memory handling.
When allocating memory for char* arrays or converting to float,
it is possible that some sort of memory error will occur. The intention
is that these events will ALWAYS raise an exception. The top-level
object conversion code now checks for errno==ENOMEM, and if so
returns NULL no matter if exceptions are requested or not.