Omit pre-validating the correctness of input for floats.
When using Python's built-in string->float converters, fastnumbers
used to pre-validate that the input was OK in order to avoid the
python built-in converter from setting an exception. The rational was
that exception setting takes a long time.

There were a couple of invalid assumptions with that logic.
 - Setting an exception does not take that long, and it's certainly
   not offset by the amount of time it takes to pre-validate.
 - The float built-in can be made to not raise an exception, which
   makes pre-validation a complete waste of time.

This is still done for integers because there is no way to disable
exception generation for PyLong_FromString, and generating an exception
actually does take a long time for that function because of an
unnecessary creation of a PyUnicode object just to build the exception.