Reformat changelog and place at root of project
Previously, the changelog lived only in the documentation. It is now in
the root of the project, and the documentation points to that file to
incorporate the changelog.

The changelog is written in markdown and conforms to the style outlined
in https://keepachangelog.com/.

Associated infrastructure files have been updated, such as Sphinx
conf.py and the tox.ini, and a requirements.txt has been added to the
docs folder so the correct markdown to rst converter can be installed.

Last, the bumpversion infrastructure has been updated for the new
changelog format.