Add switch to toggle underscore parsing in numbers
In Python >= 3.6, fastnumbers treats underscores in numbers the same way
as python does, which is to allow them if the follow a certain set of
rules. It has been requested that this behavior be toggle-able. Here,
the function that does the conversion from a Python string object to a
character array now has a boolean to disable special handling of
underscores.