Add a deploy stage
Deployment happens in four steps

1 - Source distribution
2 - Linux wheels
3 - MacOS wheels
4 - Windows wheels

All wheel building is done through a bash script that invokes the wonder
tool "cibuildwheel", which does the job of biulding the wheels on in the
various python versions. It is then uploaded with twine.

The source distribution is created with Travis's native deployment
mechanism.

In both cases, the collateral is uploaded to testing PyPI for non-tagged
commits, and to the real PyPI for tagged commits. It also is
auto-filtered from running on non-master branches and if it is not on
the main repository.