Consolidated all builtin tests
Previously, there was a different file per python version for the
builtin tests in an effort to avoid conditional logic in the tests. This
created a situation with more code to maintain than is waranted.

Now, all the tests are in one file per builtin, with conditional logic
to execute or not execute certain tests on certain python versions.