Add objects.pyx
Currently, only code to query an object type or see if an object is a
number is implemented. The functions have been removed from the C code.

Cython does not natively allow multiple .pyx files to comprise an .so
file, so to workaround this we include the non-main .pyx files into
main.pyx, rather than import them. It's a bit hacky, but it gets the job
done.