Add NumericAnalyzer C++ class
This class is designed to accept python objects that are numeric, and
determine the numeric type and how to best convert them if conversion is
needed.

This class is implemented as header-only since it is pretty small.

A Payload class is also included in this commit, but not yet used.

The NumericAnalyzer has been hooked into the query_types function - it
will be used in other function in future commits.