Potentially fixed intermittent segfault
If a length was given to vector::reserve is too large, it caused a
segfault. std::size_t was being used for the return of
PyUnicode_GET_LEGNTH instead of Py_ssize_t, and these have a sign
difference. The working hypothesis is there was an overflow happening
that caused a very large number to be passed to vector::reserve.

Using the correct types will hopefully avoid segfaults going forward.