Fix warnings and errors on MSVC
It turns out that MSVC doesn't support 'and' and 'or' for '&&' and '||',
even though it is allowed by the standard...

Also, /Wall is WAY more picky than other compilers.

- Fix signed/unsigned mismatch in assignments
- Fix potential loss of precision due to mismatch in assignments
- Fix unreferenced function parameters
- Fix conflicting argument and member variables (by prefixing members
  with 'm_')
- Fix switch statements that do not exhaustively use all enum values

One good thing is it warned that the previous preprocessor scheme was
not working as expected, so this has been fixed.