Improve performance of long string conversion
In the previous C implementation, if a string was too long to convert
naively the character array was passed to a python conversion
funciotn.

Prior to this commit in the C++ implemenation, if the character
array was too long then the original object was given to the python
conversion function. This was a waste of time because the character
array then had to be re-extracted under the hood.

Now, the extracted character array is used as it was in the C
implementation.

Furthermore, the code has actually been simplified because now we
can return Python objects in the Payload, rather than sending an action
signal indicating how to treat the python object. This removes a lot of
unnecessary and ugly code.