Use templates for the Parser in the Evaluator
Instead of storing a pointer so that we can use polymorphism, we are
using templates and a reference to the Parser in the Evaluator class.
Ideally, this will be slightly more performant because the compiler will
no longer need to make indirect calls.

As part of this, some of the methods of the Parser base class have been
made pure virutal since there is now no chance of instantiating a
Parser.