Implement old API in terms of new options system
A new options system will be used going forward to enable a cleaner API.
The existing API is implemented in terms of this options system.

A major change is the creation of the Resolver class which implements
most of the logic that used to be in the Payload class.

Additionally, some backwards-compatibility transforms have been added to
adapt old-to-new APIs.