Fix allocation error in unicode parsing function
If a string contained only spaces (at least one being a non-ASCII space)
then the logic to skip over whitespace would would completely run over
the string twice, which would then make the length be calculated as
negative. When this negative length was cast to an unsigned length, then
the array size would be too big to allocate and there would be an
exception.

Now, we protect against trying to find spaces at the end of a fully
space-only string.