Simplify type checking
Logic for determining what type is what is now simpified by having the
calling fucntion perform this analysis by examining the bit flags
created by the evaluator's type checker.

This enables lots of internal logic to be removed from evaluator.

Additionally, a bunch of helper code at the top-level was able to be
removed.

Such a strong case for "choose your abstractions wisely".