Be less defensive when parsing
The previous implementation had as its first priority not invoking
the Python exception mechanism, as this is a bit expensive, especially
if one is just going to unset it anyway. This optimized the "fail" path
at the expense of the numeric path.

The logic has been re-thought to optimize for the "happy" path first,
and assume that an acceptable trade of for very fast conversions is to
have the "fail" path take a little bit longer than before.

Scope of the changes
- Remove "is_likely" and "might_overflow" pre-checks as they are now no
  longer needed
- Remove the Parser's as_int and as_float methods - we always now just
  return Python objects
- Remove Python's string to double conversion since the C++ parser is
  now robust enough
- Use the integer parser to examine the validity of a string, instead of
  using a separate full string checker
- Add some "peek" methods to the Parser to see if the contained value is
  of a certain type without having to do a full inspection

NOTE: THERE IS A LIE ABOVE!!

This commit does not change the C++ string-to-double parser to the more
robust version, so many tests fail. The next commit will implement that
change.