Remove unneded helper functions
After re-writing parse_int and parse_float, a lot of the helpers that
had existed are no longer needed. The parsing file is now cleaned up.

As a plus, it was found that we were not being safe about parsing, and
we have to check we are not at the end of the char array when we
evaluate a character. So now parsing is safer.