Use METH_FASTCALL for all functions
This uses the "vector call" functionality to call C functions. It
requires fewer memory allocations by Python to call the function and
thus is a bit faster. Unsurprisingly, this gave a 1-2 microsecond
speedup across the board in the performance tests I ran.

Unfortunately, Python does not ship a parser for this, so I adapted the
parser from the numpy project.

This closes #59.