Add as_number template to all Parser subclasses
This method is analogous to as_pyint/as_pyfloat except that it returns
the numbers as a C type (as specified by the type given as a template
argument).

When compared to as_pyfloat there is not much difference (in fact,
CharacterParser::as_pyfloat() is implemented in terms of
CharacterParser::as_number<double>()). The error handling is identical
between the two. The only difference is there is no force_int or coerce
options.

There is a big difference between as_pyint and as_number, namely that
overflow needs to be considered for the latter. Python longs can be
contain an arbitrary number of digits, but the C integral types are
limited by the number of bytes that define them. So, e.g.
as_number<long>() needs an additional overflow error handling that
as_pyint() does not.