Modernize type hints
Utilize the bitwise-or operator instead of Union, use list instead
of List, and don't bother specifying the value of an optional argument,
just that it exists.

Unfortunately the tests need to continue to use the older style because
the runtime sees those hints and we need to support older Python
versions.