Eliminate a source of a seg fault
The NumericParser would increment the reference count of the object
on construction, and decrement on descruction. However, the
copy/assignment operators were set to default. So, on a copy operation
the reference count would be decrefed, but there would be no
corresponding incref in the copy. When that copy went out of scope, the
object would be decref again, incorrectly, creating a source of a
segfault whenever the garbage collector decided to reap the object.

The solution is to not bother with reference counts for purely internal
logic. There are of course other solutions, but this was the easiest.