Make use of std::variant for parser extraction
Instead of making the user ask which type of parser is valid and then
relying on calling that parser, we now return a std::variant containing
all possible parser types. This then simplifies using the extractor,
since now std::visit can be applied which removes a bunch of if/else
blocks.

In doing so, we no longer need the TextExtractor class, and a function
will do instead.