Simplify implementation implementataion
This is a huge commit... sorry.

This commit makes an Implementation class that can manage all the
options instead of several different functions for implementation. This
causes a net reduction in lines of code, and also makes it so we can do
thinks like keep track of reference counts (the real reason we
implemented this, but will come in a future commit).

Some other classes/logic had to be updated to support this change, but
it is worth it.