Reduce overhead for iteration
As noted in an earlier commit, there was extra work being done for the
try_* functions in the lambda. This is now resolved - the Implementation
object is created and populated, and then it is moved inside the scope
of the lambda when the lambda is created. This ensures the
Implementation object is only created once, but it stays in memory till
the iterator is finished.

To get this to work, manual move constructors had to be created that do
some reference counting.