Replace string_contains_what() with a class
Upcoming changes will require inspecting components of the string.
string_contains_what was already breaking the string into components to
do the analysis, so this has been turned into a class so that the
components could be used and analyzed layer.

As part of the process, we have used a proper enum to label the
contained types instead of just an integer.